+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Change Log
 
-## [1.1.0]
+## [1.1.1]
+- Adds autocomplete information for new BASE routines:
+  - scrDrawSpriteClipped(xbyte, y)
+  - scrDrawSpriteClippedXOR(xbyte, y)
+  - scrSetClippingView(xbyte0, y0, xbyte1, y1)
+  - scrByteToX(xbyte)
+  - scrXToByte(y)
+  - scrLineToY(yline)
+  - scrYToLine(y)
 
+## [1.1.0]
 - Adds autocomplete information for BASIC keywords.
+- Adds autocomplete information for BASE, CPCTELERA and CPCRSLIB libraries.
 
 ## [1.0.0]
-
 - Initial release.
 - Syntax highlighting